feat(core): implement "character" type (DB type "char") and PG unlimited varchar
The name "character" was picked to be different from the regular "char" type,
while still having precedent (that being alias of "char" in postgresql).

The char length is set to default of 1, to avoid excessive memory usage,
while still providing a default, as required.

For sqlite, this type is aliased to "text", just like "string".
For mssql, "character" is aliased to "nchar", to be consistent with "string" being "nvarchar",
while "char" means the native "char" type,
to be consistent with "varchar" meaning the native "varchar".
For postgresql, the "bpchar" type is also supported as an alias to "char"/"character".

Also for postgresql, if the length of "char" or "varchar" is set to -1,
that is treated as unlimited.

The defaults of both are now properly inferred during entity generation,
and omitted when not needed.